refactor #66: OptionButton 컴포넌트 분리
